03_dbt/pyproject.toml
[project]
name = "03-dbt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "dbt-bigquery>=1.11.1",
    "google-cloud-bigquery>=3.25.0",
    "google-cloud-storage>=2.16.0",
    "reportlab>=4.0.0"
]